feature/uri-self-desc: smaller functions
